runner: reopen dev right away if we know we had the lock and lost it
If we know we had the lock and it was taken, then there is no need to
try and call lock(). Just reopen the device.

For the normal case, this prevents an extra lock call.

There is a also a bug in the older rbd/ceph code where it might not
clean up internally and so the message:

 Could not break lock from 1.1.1.15:0/2175584176. (Err -16)

is repeated over and over (GH issue:
https://github.com/open-iscsi/tcmu-runner/issues/366).

This patch also prevents that problem, because the close/open
sequence before trying to grab the lock will clear the old state.

Signed-off-by: Mike Christie <mchristi@redhat.com>