rbd: increase xcopy IO size
We do not want to export that we prefer object size IOs to the initiator
because object size is normally in MBs and that ends up killing latency
due to iscsi having to break up large commads into multiple iscsi pdus.
Internally rbd is really good at larger IOs when there is no iscsi, so
tell the emulate xcopy support that we want IOs a multiple of the object
size.

Signed-off-by: Mike Christie <mchristi@redhat.com>