Add a new file_dhsmr.c device handler
Introducing a new handler, file_dhsmr.c, that emulates Zone Domains
storage devices.

This release is based on Zone Domains rev.5 draft specification.

For this initial release, only Zone Realms devices are supported,
the work is ongoing to support ZD devices without Realms support.

This version of the emulator uses an ad-hoc version of MUTATE
mechanism. A proper support for MUTATE will be added when a stable
draft of this specification becomes available.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>
Signed-off-by: David Butterfield <David.Butterfield@wdc.com>